Fixed a bug in which songs title with dash weren't able to be found
I added a function to apply the same treatment on the song title AND the guess, to be sure that name who matches the full song name will match the short one.

I added precision too, so only songs with a space before and after the dash will have this dash and what follows removed, since it's a character frequently used in few languages (like French) in other contexts, and songs with name precision after the dash frequently have the trailing spaces around dash.